% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getyear.R
\name{getyear}
\alias{getyear}
\title{Download da TISS anual por Muncípio}
\usage{
getyear(tipo, ano, uf, cod_IBGE)
}
\description{
Faz o download e extração dos dados da base TISS do FTP da ANS e retorna dados do município escolhido, procedimentos realizados, quantidade e valor no ano de interesse
}
\details{
tipo: AMBULATORIAL / HOSPITALAR

uf: unidade federativa

cod_IBGE: inteiro para código IBGE do município de 6 dígitos

ano: a partir de 2015
}
\examples{
#Loading package
library(TissLCA)
data <- getyear(AMBULATORIAL, 2019, AC, 120040)
data <- getyear(HOSPITALAR, 2019, AC, 120040)
}
