% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmonth.R
\name{getmonth}
\alias{getmonth}
\title{Download da TISS mensal por estado}
\usage{
getmonth(tipo, mes, ano, uf)
}
\description{
Faz o download e extração dos dados da base TISS do FTP da ANS e retorna dados do estado escolhido, procedimentos realizados, quantidade e valor no ano de interesse
}
\details{
tipo: AMBULATORIAL / HOSPITALAR

mes: string para o mês de interesse

uf: unidade federativa

ano: a partir de 2015
}
\examples{
# Loading package
library(TissLCA)
data <- getmonth(AMBULATORIAL, 12, 2019, AC)
}
